Création des droits d'accès avec Pundit

+++ app/controllers/candidates_controller.rb
@@ -2,21 +2,25 @@ class CandidatesController < ApplicationController
   before_action :set_candidate, only: [ :show, :edit, :update, :destroy ]
 
   def index
+    authorize Candidate
     @q = params[:q].to_s.strip
     @status = params[:status].to_s.strip
-    scope = Candidate.order(:last_name, :first_name).search(@q).with_status(@status)
+    scope = policy_scope(Candidate).order(:last_name, :first_name).search(@q).with_status(@status)
     @candidates = paginate(scope)
   end
 
   def show
+    authorize @candidate
   end
 
   def new
     @candidate = Candidate.new
+    authorize @candidate
   end
 
   def create
     @candidate = Candidate.new(candidate_params)
+    authorize @candidate
 
     if @candidate.save
       redirect_to @candidate, notice: "Candidat cree avec succes."
@@ -26,9 +30,12 @@ class CandidatesController < ApplicationController
   end
 
   def edit
+    authorize @candidate
   end
 
   def update
+    authorize @candidate
+
     if @candidate.update(candidate_params)
       redirect_to @candidate, notice: "Candidat mis a jour avec succes."
     else
@@ -37,6 +44,8 @@ class CandidatesController < ApplicationController
   end
 
   def destroy
+    authorize @candidate
+
     if @candidate.destroy
       redirect_to candidates_path, status: :see_other, notice: "Candidat supprime avec succes."
     else
