+++ app/controllers/freelancer_profiles_controller.rb
@@ -3,24 +3,28 @@ class FreelancerProfilesController < ApplicationController
   before_action :set_form_collections, only: [ :new, :create, :edit, :update ]
 
   def index
+    authorize FreelancerProfile
     @q = params[:q].to_s.strip
     @operational_status = params[:operational_status].to_s.strip
-    scope = FreelancerProfile.includes(:user, :region, :specialty)
-                             .order(created_at: :desc)
-                             .search(@q)
-                             .with_operational_status(@operational_status)
+    scope = policy_scope(FreelancerProfile).includes(:user, :region, :specialty)
+                                           .order(created_at: :desc)
+                                           .search(@q)
+                                           .with_operational_status(@operational_status)
     @freelancer_profiles = paginate(scope)
   end
 
   def show
+    authorize @freelancer_profile
   end
 
   def new
     @freelancer_profile = FreelancerProfile.new
+    authorize @freelancer_profile
   end
 
   def create
     @freelancer_profile = FreelancerProfile.new(freelancer_profile_params)
+    authorize @freelancer_profile
 
     if @freelancer_profile.save
       redirect_to @freelancer_profile, notice: "Profil freelancer cree avec succes."
@@ -30,9 +34,12 @@ class FreelancerProfilesController < ApplicationController
   end
 
   def edit
+    authorize @freelancer_profile
   end
 
   def update
+    authorize @freelancer_profile
+
     if @freelancer_profile.update(freelancer_profile_params)
       redirect_to @freelancer_profile, notice: "Profil freelancer mis a jour avec succes."
     else
@@ -41,6 +48,8 @@ class FreelancerProfilesController < ApplicationController
   end
 
   def destroy
+    authorize @freelancer_profile
+
     if @freelancer_profile.destroy
       redirect_to freelancer_profiles_path, status: :see_other, notice: "Profil freelancer supprime avec succes."
     else
@@ -55,23 +64,41 @@ class FreelancerProfilesController < ApplicationController
   end
 
   def set_form_collections
-    @users = User.order(:last_name, :first_name)
+    @users = if current_user.role_admin?
+      User.order(:last_name, :first_name)
+    else
+      User.where(id: current_user.id)
+    end
     @regions = Region.order(:name)
     @specialties = Specialty.order(:name)
   end
 
   def freelancer_profile_params
-    params.require(:freelancer_profile).permit(
-      :user_id,
-      :region_id,
-      :specialty_id,
-      :operational_status,
-      :availability_status,
-      :bio,
-      :linkedin_url,
-      :website_url,
-      :rivyr_score_current,
-      :profile_private
-    )
+    if current_user.role_admin?
+      params.require(:freelancer_profile).permit(
+        :user_id,
+        :region_id,
+        :specialty_id,
+        :operational_status,
+        :availability_status,
+        :bio,
+        :linkedin_url,
+        :website_url,
+        :rivyr_score_current,
+        :profile_private
+      )
+    else
+      params.require(:freelancer_profile).permit(
+        :region_id,
+        :specialty_id,
+        :operational_status,
+        :availability_status,
+        :bio,
+        :linkedin_url,
+        :website_url,
+        :rivyr_score_current,
+        :profile_private
+      )
+    end
   end
 end
